+++ Sfera.java
@@ -7,10 +7,10 @@ import java.awt.Graphics;
 public class Sfera extends Abs
 {
 
-  private int  _r;
+  private double  _r;
   private Color colore;
 
-  public int getR()
+  public double getR()
   {
     return _r;
   }
@@ -18,24 +18,24 @@ public class Sfera extends Abs
   public Sfera()
   {
     super();
-    _x = (int)(Math.random() * 1080);
-    _y = (int)(Math.random() * 720);
-    _r = (int)(Math.random() * 140 + 10);
-    _m = (int)(Math.random() * (1000 - 100) + 100);
+    _x = (Math.random() * 1080);
+    _y = (Math.random() * 720);
+    _r = (Math.random() * 140 + 10);
+    _m = (Math.random() * (1000 - 100) + 100);
     double col = ((double)_m-100)/(900);
-    colore = new Color((int)col,0,(int)(1-col));
+    colore = new Color((float)col,0f,(float)(1-col));
   }
 
   @Override
   protected void paintComponent(Graphics g)
   {
     g.setColor(colore);
-    g.fillOval((int)_x,(int)_y,_r,_r);
+    g.fillOval((int)_x,(int)_y,(int)_r,(int)_r);
   }
 
   @Override
   public Dimension getPreferredSize()
   {
-    return new Dimension(_r,_r);
+    return new Dimension((int)_r,(int)_r);
   }
 }
